Fix pickling article body converters
